Switched JSX files to index
