+++ src/components/LeadPhotoContainer/index.jsx
@@ -3,7 +3,7 @@
 import React, {Component} from 'react';
 
 import LeadPhoto from '../LeadPhoto/LeadPhoto';
-import PhotoInfo from '../PhotoInfo/PhotoInfo';
+import PhotoInfo from '../PhotoInfo';
 
 import './LeadPhotoContainer.css';
