--- src/components/PullQuote/PullQuote.jsx
@@ -1,21 +1,0 @@
-// @flow
-
-import React, {Component} from 'react';
-
-import './PullQuote.css';
-
-class PullQuote extends Component {
-	props: {
-		quote: string
-	};
-
-	render(){
-		return(
-			<div className='PullQuoteContainer'>
-				<p className='PullQuote'>"{this.props.quote}"</p>
-			</div>
-		)
-	}
-}
-
-export default PullQuote;
