+++ src/components/SectionHeading/index.jsx
