+++ src/components/StoryMetaDesktop/index.jsx
@@ -2,6 +2,7 @@ import React, {Component} from 'react';
 
 import Headline from '../Headline';
 import Subhead from '../Subhead';
+import StorySeriesBlock from '../StorySeriesBlock';
 import BylineContainer from '../BylineContainer';
 
 import './StoryMetaDesktop.css';
@@ -14,6 +15,7 @@ class StoryMetaDesktop extends Component{
 	render(){
 		return(
 			<div className='StoryMetaDesktop'>
+			<StorySeriesBlock text='Story Series Here'/>
 				<Headline text={this.props.meta.headline}/>
 				<Subhead text={this.props.meta.subhead}/>
 				<BylineContainer byline={this.props.meta.byline} date={this.props.meta.date}/>
