+++ src/story.txt
@@ -347,8 +347,9 @@ More events await Bassey and the Yes II Success team on summer AAU circuits, whi
 
 Bassey hasn’t received a scholarship offer from Incarnate Word. Jurenovich, the Incarnate Word vice president, said it would be unrealistic to expect a player of Bassey’s caliber to end up with the Cardinals.
 
+[.credits]
 credit: adam.zuvanich@express-news.net
-
 credit: Twitter: @AZuvanich
+[]
 
 []
